Move get_clusters_table tests from thresholding to reporting.
Also, I’ve made the subpeak detection explicitly work around binary
clusters. Now it just reports the center of mass.
